src/main/webapp/app/entities/c-country/c-country.component.ts
import { mixins } from 'vue-class-component';

import { Component, Inject } from 'vue-property-decorator';
import Vue2Filters from 'vue2-filters';
import { ICCountry } from '@/shared/model/c-country.model';
import AlertMixin from '@/shared/alert/alert.mixin';

import CCountryService from './c-country.service';

@Component
export default class CCountry extends mixins(Vue2Filters.mixin, AlertMixin) {
  @Inject('cCountryService') private cCountryService: () => CCountryService;
  private removeId: number = null;
  public itemsPerPage = 20;
  public queryCount: number = null;
  public page = 1;
  public previousPage = 1;
  public propOrder = 'id';
  public reverse = false;
  public totalItems = 0;

  public cCountries: ICCountry[] = [];

  public isFetching = false;

  public mounted(): void {
    this.retrieveAllCCountrys();
  }

  public clear(): void {
    this.page = 1;
    this.retrieveAllCCountrys();
  }

  public retrieveAllCCountrys(): void {
    this.isFetching = true;

    const paginationQuery = {
      page: this.page - 1,
      size: this.itemsPerPage,
      sort: this.sort()
    };
    this.cCountryService()
      .retrieve(paginationQuery)
      .then(
        res => {
          this.cCountries = res.data;
          this.totalItems = Number(res.headers['x-total-count']);
          this.queryCount = this.totalItems;
          this.isFetching = false;
        },
        err => {
          this.isFetching = false;
        }
      );
  }

  public prepareRemove(instance: ICCountry): void {
    this.removeId = instance.id;
    if (<any>this.$refs.removeEntity) {
      (<any>this.$refs.removeEntity).show();
    }
  }

  public removeCCountry(): void {
    this.cCountryService()
      .delete(this.removeId)
      .then(() => {
        const message = this.$t('opusWebApp.cCountry.deleted', { param: this.removeId });
        this.alertService().showAlert(message, 'danger');
        this.getAlertFromStore();
        this.removeId = null;
        this.retrieveAllCCountrys();
        this.closeDialog();
      });
  }

  public sort(): Array<any> {
    const result = [this.propOrder + ',' + (this.reverse ? 'asc' : 'desc')];
    if (this.propOrder !== 'id') {
      result.push('id');
    }
    return result;
  }

  public loadPage(page: number): void {
    if (page !== this.previousPage) {
      this.previousPage = page;
      this.transition();
    }
  }

  public transition(): void {
    this.retrieveAllCCountrys();
  }

  public changeOrder(propOrder): void {
    this.propOrder = propOrder;
    this.reverse = !this.reverse;
    this.transition();
  }

  public closeDialog(): void {
    (<any>this.$refs.removeEntity).hide();
  }
}
